backend/requirements_test.txt
pytest
ipdb
sanic-testing
pytest-asyncio
peewee